Updated code according to MVC architecture
