--- JS/modules/imageName.js
@@ -1,7 +1,0 @@
-export default function createName (imgID) {
-    var imagename = document.createElement("p")
-
-    imagename.setAttribute('class', 'imagename')
-
-    return imagename;
-}
